requirements.txt
dj-database-url==0.4.2
Django==2.0.2
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
mysqlclient==1.3.12
psycopg2-binary==2.7.4
PyJWT==1.5.3
pytz==2017.3
whitenoise==3.3.1
